dashboard/pyproject.toml
[project]
name = "dashboard"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pyserial>=3.5",
    "reaktiv>=0.19.3",
]
